requirements.txt
asgiref==3.7.2
cffi==1.15.1
distlib==0.3.7
Django==4.2.5
django-shortcuts==1.6
djangorestframework==3.14.0
docutils==0.20.1
filelock==3.12.3
platformdirs==3.10.0
pycparser==2.21
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
virtualenv==20.24.4
